requirements.txt
# Core dependencies
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# Development tools
django-debug-toolbar==4.4.6
black==24.10.0
flake8==7.1.1
mypy-extensions==1.0.0
pycodestyle==2.12.1
pyflakes==3.2.0
mccabe==0.7.0

# Utilities and libraries
attrs==24.2.0
click==8.1.7
Markdown==3.7
packaging==24.1
platformdirs==4.3.6
python-dotenv==1.0.1
PyYAML==6.0.2
uritemplate==4.1.1
sqlparse==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
referencing==0.35.1
inflection==0.5.1
rpds-py==0.20.0
psycopg~=3.2.3
